requirements.txt
azure-functions>=1.12.0
fastapi
httpx
pydantic